% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{addEnsembleToTS}
\alias{addEnsembleToTS}
\title{Ensemble Timeseries Functions}
\usage{
addEnsembleToTS(
  ensembleTS,
  issueDate,
  startDate,
  interval,
  ensembleArray,
  units = NULL
)
}
\arguments{
\item{ensembleTS}{ensemble timeseries to add ensemble to}

\item{issueDate}{date for which ensemble should be added}

\item{startDate}{date on which ensemble values starts}

\item{interval}{duration object representing interval between values in forecast}

\item{ensembleArray}{data}
}
\value{
void
}
\description{
Ensemble Timeseries Functions
}
