% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{closeEnsemblesFile}
\alias{closeEnsemblesFile}
\title{Ensemble file handling functions}
\usage{
closeEnsemblesFile(ensemblesFile)
}
\arguments{
\item{ensemblesFile}{java object}
}
\description{
Ensemble file handling functions
}
