% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{deleteRecord}
\alias{deleteRecord}
\title{Manages records in a file}
\usage{
deleteRecord(filename, location, parameter)
}
\arguments{
\item{parameter}{}
}
\value{
true if successful
}
\description{
Deletes records from within a file
}
