% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{.jTimestampToPOSIXct}
\alias{.jTimestampToPOSIXct}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
.jTimestampToPOSIXct(jTimestamp)
}
\arguments{
\item{jTimestamp}{}
}
\value{
POSIXct from java ZonedDateTime object
}
\description{
Functions to assist with TSEnsembles timestamps
}
