% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{.jTimezone}
\alias{.jTimezone}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
.jTimezone(tzCode)
}
\arguments{
\item{tzCode}{}
}
\value{
java object for timezone
}
\description{
Functions to assist with TSEnsembles timestamps
}
