% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{durationDays}
\alias{durationDays}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
durationDays(days)
}
\arguments{
\item{days}{}
}
\value{
Java Duration object
}
\description{
Functions to assist with TSEnsembles timestamps
}
