% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{durationHours}
\alias{durationHours}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
durationHours(hours)
}
\arguments{
\item{hours}{}
}
\value{
Java Duration object
}
\description{
Functions to assist with TSEnsembles timestamps
}
