% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{durationMinutes}
\alias{durationMinutes}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
durationMinutes(minutes)
}
\arguments{
\item{minutes}{}
}
\value{
Java Duration object
}
\description{
Functions to assist with TSEnsembles timestamps
}
