% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{getCatalog}
\alias{getCatalog}
\title{Ensemble file handling functions}
\usage{
getCatalog(ensemblesFile)
}
\arguments{
\item{ensemblesFile}{}
}
\value{
data.frame of file catalog
}
\description{
Ensemble file handling functions
}
