% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{getEnsembleAtTime}
\alias{getEnsembleAtTime}
\title{Ensemble Timeseries Functions}
\usage{
getEnsembleAtTime(ensembleTS, issueDate, toMatrix = TRUE)
}
\arguments{
\item{timestamp}{}
}
\description{
Ensemble Timeseries Functions
}
