% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getRecords}
\alias{getRecords}
\title{Manage records in file}
\usage{
getRecords(filename)
}
\arguments{
\item{filename}{}
}
\value{
data.frame of locations in file
}
\description{
Manage records in file
}
