% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{getTidyEnsembleAtTime}
\alias{getTidyEnsembleAtTime}
\title{Ensemble Timeseries Functions}
\usage{
getTidyEnsembleAtTime(
  ensembleTS,
  issueDate,
  paramLabel = NULL,
  addMetadata = TRUE
)
}
\arguments{
\item{addMetadata}{}
}
\value{
a tidy ensemble for \code{issueDate}
}
\description{
Ensemble Timeseries Functions
}
