% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{getTidyEnsembles}
\alias{getTidyEnsembles}
\title{Ensemble Timeseries Functions}
\usage{
getTidyEnsembles(ensembleTS)
}
\arguments{
\item{ensembleTS}{}
}
\value{
tidy data.frame of ensembles for all issue dates in ensembleTS
}
\description{
Ensemble Timeseries Functions
}
