% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{newEnsembleTS}
\alias{newEnsembleTS}
\title{Ensemble Timeseries Functions}
\usage{
newEnsembleTS(record, units, type, version)
}
\arguments{
\item{version}{}
}
\description{
Ensemble Timeseries Functions
}
