% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{newTimestamp}
\alias{newTimestamp}
\title{Functions to assist with TSEnsembles timestamps}
\usage{
newTimestamp(
  tsYear,
  tsMonth,
  tsDay,
  tsHour = 0,
  tsSeconds = 0,
  tzID = .jTimezone("Z")
)
}
\arguments{
\item{tzID}{}
}
\value{
timestamp as ZoneDateTime object
}
\description{
Functions to assist with TSEnsembles timestamps
}
