% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{openEnsemblesFile}
\alias{openEnsemblesFile}
\title{Ensemble file handling functions}
\usage{
openEnsemblesFile(filename, type = "SQLite", mode = "append")
}
\arguments{
\item{type}{"SQLite" (default) or "DSS" to determine what type of file to open}

\item{create.mode}{"append" (default), "new" which prevents opening an existing file, or "overwrite" which deletes existing and creates a new file}
}
\value{
Java object for ensemble file
}
\description{
Ensemble file handling functions
}
