% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{toEnsembleArray}
\alias{toEnsembleArray}
\title{Ensemble Timeseries Functions}
\usage{
toEnsembleArray(tidyEnsemble)
}
\arguments{
\item{tidyEnsemble}{}
}
\value{
an array in format stored inside of TSEnsembles, can be written to file
}
\description{
Ensemble Timeseries Functions
}
