% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{toTidyEnsemble}
\alias{toTidyEnsemble}
\title{Ensemble Timeseries Functions}
\usage{
toTidyEnsemble(arrayEnsemble, startTime, interval, paramLabel)
}
\arguments{
\item{paramLabel}{}
}
\value{
tidy data.frame format containing the ensemble array with timestamps
}
\description{
Ensemble Timeseries Functions
}
