% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{writeToEnsemblesFile}
\alias{writeToEnsemblesFile}
\title{Ensemble file handling functions}
\usage{
writeToEnsemblesFile(ensemblesFile, object)
}
\arguments{
\item{ensemblesFile}{}
}
\description{
Ensemble file handling functions
}
